+++ MyFitnessJournal Watch App/Features/B_Test/Manager/DefaultTrainingManager.swift
@@ -9,6 +9,24 @@ import HealthKit
 
 final class DefaultTrainingManager: NSObject, TrainingManager {
     
+    let shareTypes: Set<HKSampleType> = [
+        HKQuantityType.workoutType()
+    ]
+    
+    let readTypes: Set<HKObjectType> = [
+        HKQuantityType(.heartRate),
+        HKQuantityType(.activeEnergyBurned),
+        HKQuantityType(.workoutEffortScore),
+        HKObjectType.activitySummaryType()
+    ]
+    
+    
+    func requestAuthorization() {
+        healthStore.requestAuthorization(toShare: shareTypes, read: readTypes) { (success, error) in
+            // Handle error.
+        }
+    }
+    
     let healthStore = HKHealthStore()
     
     var session: HKWorkoutSession?
