+++ MyFitnessJournal Watch App/Features/B_Test/Manager/TrainingManager.swift
@@ -6,43 +6,50 @@
 //
 
 import HealthKit
-import HealthKit
 
-/// A protocol that defines an interface for managing a HealthKit-powered workout session.
+/// A protocol that manages HealthKit-based workout sessions.
+///
+/// This includes starting a workout, toggling pause/resume, ending the session,
+/// and exposing real-time streams for workout metrics and session state.
 protocol TrainingManager {
     
-    // MARK: - Configuration
+    /// The current HKLiveWorkoutBuilder instance used to collect live workout data.
+    ///
+    /// This is exposed for internal use and should not be modified externally.
+    var builder: HKLiveWorkoutBuilder? { get }
     
-    /// The currently selected workout activity type (e.g. walking, running).
-    var selectedWorkout: HKWorkoutActivityType? { get set }
+    // MARK: - Authorization
     
-    /// Sets the selected workout type.
+    func requestAuthorization()
+    
+    // MARK: - Workout Configuration
+    
+    /// Sets the workout type and initializes a new HealthKit session.
+    ///
+    /// - Parameter type: The workout activity type to be tracked.
     func setSelectedWorkout(_ type: HKWorkoutActivityType?)
     
-    /// The current workout builder instance.
-    var builder: HKLiveWorkoutBuilder? { get }
-
-    // MARK: - Workout State
-    
-    /// A boolean indicating whether the workout session is running.
-    var workoutSessionIsRunning: Bool { get }
-    
-    /// A stream that emits the current `workoutSessionIsRunning` state when it changes.
-    func workoutSessionIsRunningStream() -> AsyncStream<Bool>
-    
-    // MARK: - Workout Lifecycle Control
-    
+    /// Sets the state of the summary view (shown at the end of a workout).
+    ///
+    /// - Parameter value: `true` to show the summary view, `false` to hide it.
     func setValueForSumaryView(_ value: Bool)
-//    var showingSummaryView: Bool { get set }
-
-    /// Toggles pause/resume state of the workout session.
+    
+    // MARK: - Session Lifecycle
+    
+    /// Pauses or resumes the workout session depending on its current state.
     func togglePause()
     
     /// Ends the current workout session.
     func endWorkout()
     
-    // MARK: - Workout Metrics
+    // MARK: - Streams
     
-    /// A stream that emits updated workout metrics during the session.
+    /// A stream that emits live workout metrics, such as heart rate and calories burned.
     var workoutMetricsStream: AsyncStream<WorkoutMetrics> { get }
+    
+    /// A stream that emits the running state of the workout session.
+    ///
+    /// Emits `true` if the workout is currently running, otherwise `false`.
+    func workoutSessionIsRunningStream() -> AsyncStream<Bool>
+    
 }
