+++ MyFitnessJournal Watch App/Features/B_Test/Summary/SubView/SummaryMetricView.swift
@@ -2,7 +2,7 @@
 //  SummaryMetricView.swift
 //  StepTrackerTCA
 //
-//  Created by Sebastian Sciuba on 23/05/2025.
+//  Created by Sebastian Sciuba on 31/05/2025.
 //
 
 import SwiftUI
