+++ MyFitnessJournal Watch App/Features/Main/MainViewAW.swift
@@ -31,32 +31,14 @@ struct MainViewAW: View {
             }
             .listStyle(.carousel)
             .navigationBarTitle("Workouts")
-            .sheet(item: $store.scope(state: \.destination?.openSummary,
-                                      action: \.destination.openSummary)) { store in
-                SummaryView(store: store)
+            .sheet(item: $store.scope(state: \.destination?.openTrainingSummary,
+                                      action: \.destination.openTrainingSummary)) { store in
+                TrainingSummaryView(store: store)
             }
-                                      .sheet(item: $store.scope(state: \.destination?.openTrainingSummary,
-                                                                action: \.destination.openTrainingSummary)) { store in
-                                          TrainingSummaryView(store: store)
-                                      }
-                                                                .navigationDestination(
-                                                                    item: $store.scope(
-                                                                        state: \.destination?.workoutSession,
-                                                                        action: \.destination.workoutSession)) { store in
-                                                                            WorkoutSessionView(store: store)
-                                                                        }
-                                                                        .navigationDestination(
-                                                                            item: $store.scope(
-                                                                                state: \.destination?.openTest,
-                                                                                action: \.destination.openTest)) { store in
-                                                                                    HeartRateView(store: store)
-                                                                                }
-                                                                                .navigationDestination(
-                                                                                    item: $store.scope(
-                                                                                        state: \.destination?.trainingSession,
-                                                                                        action: \.destination.trainingSession)) { store in
-                                                                                            TrainingSessionTabView(store: store)
-                                                                                        }
+            .navigationDestination(item: $store.scope(state: \.destination?.trainingSession,
+                                                      action: \.destination.trainingSession)) { store in
+                TrainingSessionTabView(store: store)
+            }
         }
         .onAppear {
             send(.viewDidAppear)
@@ -71,5 +53,3 @@ struct MainViewAW: View {
         MainFeatureAW()
     }))
 }
-
-
