+++ MyFitnessJournal Watch App/Features/Main/Model/WorkoutMetrics.swift
@@ -2,7 +2,7 @@
 //  WorkoutMetrics.swift
 //  StepTrackerTCA
 //
-//  Created by Sebastian Sciuba on 25/05/2025.
+//  Created by Sebastian Sciuba on 31/05/2025.
 //
 
 import Foundation
