+++ MyFitnessJournal Watch App/Features/Main/Utilities/DefaultMainServiceAW.swift
@@ -1,23 +1,32 @@
 //
-//  DefaultMainServiceAW.swift
+//  AuthorizationAW.swift
 //  MyFitnessJournal Watch App
 //
 //  Created by Sebastian Sciuba on 23/05/2025.
 //
 
-import Factory
+import ComposableArchitecture
 import Foundation
 
-final class DefaultMainServiceAW: MainServiceAW {
-    
-    // MARK: - Dependency
-    
-    @Injected(\.workoutManager) private var workoutManager
-    
-    // MARK: - API
-    
-    func requestAuthorization() {
-        workoutManager.requestAuthorization()
-    }
-    
+struct AuthorizationClientAW {
+    var requestAuthorization: @Sendable () -> Void
 }
+
+extension DependencyValues {
+    var authorizationClientAW: AuthorizationClientAW {
+        get { self[AuthorizationKeyAW.self] }
+        set { self[AuthorizationKeyAW.self] = newValue }
+    }
+}
+
+private enum AuthorizationKeyAW: DependencyKey {
+    static let liveValue: AuthorizationClientAW = {
+        
+        @Dependency(\.trainingManager) var manager
+        
+        return AuthorizationClientAW {
+            manager.requestAuthorization()
+        }
+    }()
+}
+
