+++ MyFitnessJournal Watch App/Features/Main/Utilities/MainServiceAW.swift
@@ -7,8 +7,8 @@
 
 import Foundation
 
-protocol MainServiceAW {
-    
-    func requestAuthorization()
-    
-}
+//protocol MainServiceAW {
+//    
+//    func requestAuthorization()
+//    
+//}
