--- MyFitnessJournal Watch App/Features/Workout/Summary/Utilities/SummaryService.swift
@@ -1,16 +1,0 @@
-//
-//  SummaryService.swift
-//  MyFitnessJournal Watch App
-//
-//  Created by Sebastian Sciuba on 25/05/2025.
-//
-
-import Foundation
-
-protocol SummaryService {
-    
-    ///
-    func setShowingSummaryView(_ isShowing: Bool)
-    
-}
-
