--- MyFitnessJournal Watch App/Features/Workout/WorkoutSession/Child/WorkoutMetric/Utilities/WorkoutMetricService.swift
@@ -1,18 +1,0 @@
-//
-//  WorkoutMetricService.swift
-//  MyFitnessJournal Watch App
-//
-//  Created by Sebastian Sciuba on 23/05/2025.
-//
-
-import Foundation
-import Factory
-import SwiftUI
-
-protocol WorkoutMetricService {
-    
-    var workoutMetricsStream: AsyncStream<WorkoutMetrics> { get }
-    
-//    func elapsedTime(at context: TimelineViewDefaultContext) -> TimeInterval?
-    
-}
