--- MyFitnessJournal Watch App/Service/WorkoutManager/WorkoutManager+Delegate/HKLiveWorkoutBuilderDelegate.swift
@@ -1,35 +1,0 @@
-//
-//  HKLiveWorkoutBuilderDelegate.swift
-//  StepTrackerTCA
-//
-//  Created by Sebastian Sciuba on 23/05/2025.
-//
-
-import Foundation
-import HealthKit
-
-extension DefaultWorkoutManager: HKWorkoutSessionDelegate {
-    
-    ///
-    func workoutSession(_ workoutSession: HKWorkoutSession, didChangeTo toState: HKWorkoutSessionState,
-                        from fromState: HKWorkoutSessionState, date: Date) {
-        DispatchQueue.main.async {
-            self.workoutSessionIsRunning = toState == .running
-        }
-
-        if toState == .ended {
-            builder?.endCollection(withEnd: date) { (success, error) in
-                self.builder?.finishWorkout { (workout, error) in
-                    DispatchQueue.main.async {
-                        self.workout = workout
-                    }
-                }
-            }
-        }
-        print("🔄 Workout session changed from \(fromState.rawValue) to \(toState.rawValue)")
-    }
-
-    ///
-    func workoutSession(_ workoutSession: HKWorkoutSession, didFailWithError error: Error) { }
-    
-}
